requirements.txt
# System
aiohttp
dnspython
PyNaCl
async-timeout
pyyaml
# Discord
discord
discord.py
discordUtils